refactor(api): move ScriptRecord import to db_script_ops for clarity
- Updated import of ScriptRecord in src/scriptrag/api/__init__.py to come from db_script_ops instead of database_operations.
- Removed ScriptRecord from __all__ in database_operations.py to reflect this change.
- Adjusted test imports accordingly.
- Minor comment and docstring clarifications in query loader and test fixtures for better clarity.

This refactor improves module organization and clarifies the source of ScriptRecord.
