+++ src/scriptrag/api/__init__.py
@@ -15,7 +15,7 @@ from scriptrag.api.analyze_results import AnalyzeResult as AnalyzeResult
 from scriptrag.api.analyze_results import FileResult as FileResult
 from scriptrag.api.database import DatabaseInitializer as DatabaseInitializer
 from scriptrag.api.database_operations import DatabaseOperations as DatabaseOperations
-from scriptrag.api.database_operations import ScriptRecord as ScriptRecord
+from scriptrag.api.db_script_ops import ScriptRecord as ScriptRecord
 from scriptrag.api.index import IndexCommand as IndexCommand
 from scriptrag.api.index import IndexOperationResult as IndexOperationResult
 from scriptrag.api.index import IndexResult as IndexResult
